Configure video source when VideoPlayer is mounted
Setting source when props are updates somehow didn't work in browser
(in isolation == autoreload mode, it worked when imported into web app).
And we figured out that we don't need to update source
for mounted player. So no point wasting time on it now when it is not
needed.
VTMS-2131